Resolve the proto import path for bazel and gradle.
See comments in top level BUILD.bazel.

Define the proto rule at the root level, instead of under src/main/proto,
to resolve the import root dir difference between bazel and gradle.
Specifically, proto-gradle-plugin considers src/main/proto as the root, so
the proto `import` path should be relative. Whereas bazel considers the top
level as the root, so the import should be absolute (or relative to the top).

The solution is to update the basel side by defining the proto rule at the
top level, then strip_import_prefix such that the protos are relative.

Tested with `bazel build //src/main/cpp` and `./gradlew assemble`.